refactor: routes.ts를 shared/로 이동하여 단방향 의존성 유지
- app/routes에서 shared/config/routes로 이동
- pages에서 app에 대한 의존성 제거
