+++ src/pages/cardComplete/CardComplete.tsx
@@ -1,5 +1,5 @@
 import { useNavigate, useLocation } from 'react-router-dom';
-import { ROUTES } from '../../app/routes';
+import { ROUTES } from '../../shared/config/routes';
 import { ICON_PATH } from '../../shared/constants/cardImagePaths';
 import './cardComplete.css';
